fix: updated documentation, moved BasicAuthModule to main scope * renamed BaseAuthModule to AbstractAuthModule and made class abstract
