+++ jetty-jaspi/src/main/java/org/eclipse/jetty/security/jaspi/SimpleServerAuthContext.java
@@ -23,6 +23,10 @@ import jakarta.security.auth.message.MessageInfo;
 import jakarta.security.auth.message.config.ServerAuthContext;
 import jakarta.security.auth.message.module.ServerAuthModule;
 
+/** 
+ * Simple bridge implementation of the Jakarta Authentication {@link ServerAuthContext} ifc, 
+ * which will delegate to the provided {@link ServerAuthModule} implementation. 
+ */
 public class SimpleServerAuthContext implements ServerAuthContext
 {
     private final ServerAuthModule serverAuthModule;
